+++ build_daemon/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 4.1.3-wip
 
-- Put lock file under `.dart_tool` in the workspace by default. Add
-  `daemonSharedPath` to keep support for other locations.
+- Put lock and config files under `.dart_tool` in the workspace by default.
+  Make them user private. Add `daemonSharedPath` to keep support for other
+  locations.
+- Use a token for websocket authentication.
 - Require Dart 3.11.0.
 
 ## 4.1.2